backend/requirements.txt
requests
openai
python-dotenv
pygame
python-dotenv
websockets
elevenlabs
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
deepgram-sdk
mem0ai>=0.1.0
asyncio>=3.4.0
vecs>=0.4.0
numpy>=1.24.0
scikit-learn>=1.3.0
textstat>=0.7.0
pytest-asyncio>=0.23.0
graphviz
mermaid
ipython